Create board.hpp
The `game::board` interface defines a matrix of numbers representing a
generic playing board. The user can place numbers into the board and
test for fullness or the presence of specific formations of numbers,
particularly the same number repeating K number of times in a row
horizontally, vertically or diagonally.