Update transformed.cpp
An effective unit test relies on no external dependencies to function.
utils::transformed<T> will be updated to enable passing an anonymous
list of T values as a constructor argument, without requiring a
concrete sequence<T> to be instantiated at the outermost call site. The
test was further changed to establish a more convincing expected
outcome.