Improve typing of XOBoard dimensions
The width and height attributes are used to reason about the accessable
boundaries of XOBoard's contents; as such, the attributes should be
typed appropriately for iteration and indexing of XOBoard. Presently,
the contents are stored in a vector<size_t>, so the typename
vector<size_t>::size_type gives us the correct type to use.

(Ultimately, this type still resolves as size_t; however, if changes
are made to the way in which XOBoard's contents are stored, a different
type may be used here)